feat: add background job to run calculator

+++ StakingPointsSystem/Models/Asset.cs
@@ -7,7 +7,7 @@ public class Asset
     public int UserId { get; set; }
     public TransactionType TransactionType { get; set; }
     public AssetType AssetType { get; set; }
-    public int Unit { get; set; }
+    public decimal Unit { get; set; }
     public DateTime CreatedTime { get; set; }
 
     public IStatement ToStatement()
